+++ .mocharc.json
@@ -1,6 +1,7 @@
 {
   "extension": ["ts"],
   "loader": "ts-node/esm",
-  "spec": "test/*.test.ts",
+  "spec": "./src/**/*.test.ts",
   "require": "mochaSetup.js"
 }
+
