+++ tsconfig.json
@@ -19,5 +19,6 @@
     "sourceMap": true,
     "strictNullChecks": true
   },
-  "include": ["src", "test/HTTPTransport.test.ts", "test/router.test.ts"]
+  "include": ["src"]
 }
+
